requirements.txt
fastapi
pandas
streamlit
torch
torchvision
pillow
Python-IO